Added item picklist
- added generic picklist limited to item type
- fixed scrollbar hiding top menu
- changed item cost to double to allow <1 gold cost
- added some more items to the item.yaml file

+++ src/Wayfinder.Infrastructure/DTOs/ItemYamlDto.cs
@@ -7,7 +7,11 @@ namespace Wayfinder.Infrastructure.DTOs
         public string Id { get; set; } = string.Empty;
         public string Name { get; set; } = string.Empty;
         public double Weight { get; set; }
-        public int Cost { get; set; }
+        // 1 = 1 gp
+        // 0.1 = 1 sp
+        // 0.01 = 1 cp
+        public double Cost { get; set; }
+
         [YamlMember(Alias = "Type")]
         public string ItemType { get; set; }
         public string Description { get; set; } = string.Empty;
